requirements.txt
numpy==1.16.4
tensorflow-cpu==2.2.0